pyproject.toml
[tool.poetry]
name = "moscow-transport"
version = "0.1.0"
description = ""
authors = ["Averkina Olga https://github.com/Averkina-2020"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
requests = "^2.32.3"
urllib3 = "1.26.15"
beautifulsoup4 = "^4.12.3"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
apscheduler = "^3.10.4"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
